backend/requirements.txt
fastapi
uvicorn[standard]
supabase
google-genai
httpx
python-dotenv
railtracks
